requirements.txt
streamlit
pandas
numpy
catboost
scikit-learn
torch
torchvision
Pillow
joblib